+++ packages/keyring-eth-trezor/src/trezor-keyring.ts
@@ -146,7 +146,7 @@ export class TrezorKeyring implements Keyring {
   async init(): Promise<void> {
     return this.bridge.init({
       manifest: TREZOR_CONNECT_MANIFEST,
-      lazyLoad: process.env.IN_TEST ? false : true,
+      lazyLoad: true,
     });
   }